feat(accessibility): Add `progressSemantic` for customizable progress indicator semantics
This commit introduces a new optional parameter `progressSemantic` to the `IntroductionScreen` widget. This parameter allows developers to provide a custom semantic label for the progress indicator (dots or custom progress).

If `progressSemantic` is provided, it will be used to generate the semantic label. Otherwise, a default label "Page {currentPage} of {totalPages}" will be used.

This change enhances accessibility by allowing developers to provide more contextually relevant information to users relying on screen readers.